Move Members to shared
