+++ src/components/shared/multisignatureMembers/index.js
@@ -1,0 +1,41 @@
+import React from 'react';
+
+import AccountVisual from '../../toolbox/accountVisual';
+import regex from '../../../utils/regex';
+
+import styles from './styles.css';
+
+const Member = ({ member, i, t }) => (
+  <div className={`${styles.memberInfo} member-info`}>
+    <p className={styles.memberInfoIndex}>{`${i + 1}.`}</p>
+    <AccountVisual address={member.address} />
+    <div className={styles.memberDetails}>
+      <p className={styles.memberTitle}>
+        {member.name || member.address.replace(regex.lskAddressTrunk, '$1...$3')}
+        <span>{`(${member.mandatory ? t('Mandatory') : t('Optional')})`}</span>
+      </p>
+      <p className={styles.memberKey}>{member.publicKey.replace(regex.publicKeyTrunk, '$1...$3')}</p>
+    </div>
+  </div>
+);
+
+const Members = ({ members, t }) => {
+  const sliceIndex = Math.round(members.length / 2);
+  const leftColumn = members.slice(0, sliceIndex);
+  const rightColumn = members.slice(sliceIndex, members.length);
+  return (
+    <div className={styles.membersContainer}>
+      <p>{t('Members')}</p>
+      <div>
+        {leftColumn.map((member, i) =>
+          <Member member={member} i={i} key={`registerMultiSignature-members-list-${i}`} t={t} />)}
+      </div>
+      <div>
+        {rightColumn.map((member, i) =>
+          <Member member={member} i={i + sliceIndex} key={`registerMultiSignature-members-list-${i + sliceIndex}`} t={t} />)}
+      </div>
+    </div>
+  );
+};
+
+export default Members;
